+++ exportplan/management/commands/report_helper.py
@@ -1,4 +1,5 @@
 from exportplan import models
+import csv
 
 
 def is_ep_plan_empty(plan, fields):
@@ -20,6 +21,7 @@ def is_ep_plan_empty(plan, fields):
 def set_useable_fields():
     not_needed_model_fields = [
         "id",
+        "name",
         "created",
         "modified",
         "sso_id",
@@ -30,3 +32,19 @@ def set_useable_fields():
     ]
     my_model_fields = [field.name for field in models.CompanyExportPlan._meta.get_fields()]
     return [field for field in my_model_fields if field not in not_needed_model_fields]
+
+
+def write_ep_csv(ep_list, path):
+    with open(path, mode='w') as exportplan:
+        fieldnames = ['sso_id', 'export_countries', 'export_commodity_codes']
+        writer = csv.DictWriter(exportplan, fieldnames=fieldnames)
+
+        writer.writeheader()
+        for ep in ep_list:
+            writer.writerow(
+                {
+                    'sso_id': ep["sso_id"],
+                    'export_countries': ep["export_countries"],
+                    'export_commodity_codes': ep["export_commodity_codes"],
+                }
+            )