+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - GP2-3173 - multi-Export-plan-list-detail
 - GP2-3351 - Add permission class to export-plan api can only be changed/accessed by owner 
 - GP2-3179 - Spike management command: report_export_plan
+- GP2-3646 - Saving down exportplan to csv command
 
 
 ## [1.14.0](https://github.com/uktrade/directory-api/releases/tag/1.14.0)